backup/requirements.txt
google-auth
google-auth-httplib2
google-api-python-client